backend/src/ai_model/requirements.txt
google-cloud-storage
langchain-google-genai
langchain_community
langchain-text-splitters
langchain-chroma
PyPDF2
python-dotenv
google-generativeai
google-genai
httpx